+++ backend/src/App.Domain/Dtos/HistoryTicker/HistoryTickerDtoCreateResult.cs
@@ -4,14 +4,14 @@ namespace App.Domain.Dtos.HistoryTicker
     {
         public int ArquivoImportacaoId { get; set; }
         public int TickerId { get; set; }
-        public float PrecoUnitario { get; set; }
-        public float PrecoLucro { get; set; }
-        public float Roic { get; set; }
-        public float EvEbit { get; set; }
-        public float MargemEbit { get; set; }
-        public float? DividendYield { get; set; }
-        public float? PrecoValorPatrimonial { get; set; }
-        public float? LiquidezMediaDiaria { get; set; }
-        public float? ValorMercado { get; set; }
+        public decimal PrecoUnitario { get; set; }
+        public decimal PrecoLucro { get; set; }
+        public decimal Roic { get; set; }
+        public decimal EvEbit { get; set; }
+        public decimal MargemEbit { get; set; }
+        public decimal? DividendYield { get; set; }
+        public decimal? PrecoValorPatrimonial { get; set; }
+        public decimal? LiquidezMediaDiaria { get; set; }
+        public decimal? ValorMercado { get; set; }
     }
 }
